% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Percentage Change (over 12 months) of United Kingdom Consumer Price Inflation}
\format{
Time-Series [1:67] from 1997 to 2014:
88.8, 89.6, 90, 90.4, 90.3, 91.5, 91.2, 91.7, 91.5, 92.7, ...
}
\source{
October 2013 CPI data by the Office of National Statistics.
Retrieved from "Consumer Price Inflation Reference Tables, October 2013".

Original spreadsheet no longer available on the ONS website,
but a copy exists at
\url{https://github.com/guyabel/fanplot/tree/master/data-raw/}
}
\usage{
data(cpi)
}
\description{
Time series of quarterly UK CPI from Q1 1997 to Q3 2013.
Data formatted from the October 2013 release of the CPI data by the Office of National Statistics.
Q1 values are taken from February, Q2 from May, Q3 from August, and Q4 from November.
}
\examples{
data(cpi)

}
\keyword{datasets}
