\name{fanplot-package}
\alias{fanplot-package}
\alias{fanplot}
\docType{package}
\title{
Visualisation of Sequential Probability Distributions Using Fan Charts.
}
\description{
Visualise sequential distributions using a range of plotting styles. Sequential distribution data can be input as either simulations or values corresponding to percentiles over time. Plots are added to existing graphic devices using the fan function. Users can choose from four different styles, including fan chart type plots, where a set of coloured polygon, with shadings corresponding to the percentile values are layered to represent different uncertainty levels.
}
\details{
\tabular{ll}{
Package: \tab fanplot\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
}
Blog posts with some additional details of the implementation of functions in the package can be found at
\url{http://gjabel.wordpress.com/category/r/fanplot/} 

Github repo:
\url{http://github.com/gjabel/fanplot} 

R Journal Paper:
\url{http://journal.r-project.org/archive/2015-1/abel.pdf}

}
\author{
Guy J. Abel
}
\references{
Abel, G. J. (2015). fanplot: An R Package for visualising sequential distributions. \emph{The R Journal} 7 (2) 15--23.
}

