% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanplot-package.R
\docType{package}
\name{fanplot-package}
\alias{fanplot}
\alias{fanplot-package}
\title{fanplot: Visualisation of Sequential Probability Distributions Using Fan Charts}
\description{
Visualise sequential distributions using a range of plotting styles.
Sequential distribution data can be input as either simulations or values
corresponding to percentiles over time. Plots are added to existing graphic
devices using the \code{\link{fan}} function. Users can choose from four
different styles, including fan chart type plots, where a set of coloured
polygons, with shadings corresponding to the percentile values, are layered
to represent different uncertainty levels.
}
\details{
License: GPL-2
}
\references{
Abel, G. J. (2015). fanplot: An R Package for visualising sequential distributions.
\emph{The R Journal}, 7(2), 15--23.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://guyabel.github.io/fanplot/}
  \item Report bugs at \url{https://github.com/guyabel/fanplot/issues/}
}

}
\author{
Guy J. Abel
}
