% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svpdx.R
\docType{data}
\name{svpdx}
\alias{svpdx}
\title{Pound–Dollar Exchange Rate Data}
\format{
A data frame with 945 observations on the following 2 variables:
\describe{
  \item{date}{Date of observation.}
  \item{pdx}{Logarithm of returns for Pound–Dollar exchange.}
}
}
\source{
\url{http://www.econ.vu.nl/koopman/sv/svpdx.dat}
}
\usage{
data(svpdx)
}
\description{
Pound–Dollar exchange rate data from 2 October 1981 to 28 June 1985.
}
\details{
Raw data on log returns.
}
\examples{
data(svpdx)

# plot
plot(svpdx$pdx, type = "l", xaxt = "n",
     xlab = "Time", ylab = "Return")

# add x-axis
svpdx$rdate <- format(svpdx$date, format = "\%b \%Y")
mth <- unique(svpdx$rdate)
qtr <- mth[seq(1, length(mth), 3)]
axis(1, at = match(qtr, svpdx$rdate), labels = qtr, cex.axis = 0.75)
axis(1, at = match(mth, svpdx$rdate), labels = FALSE, tcl = -0.2)

}
\references{
Meyer, R. and J. Yu (2002).
BUGS for a Bayesian analysis of stochastic volatility models.
\emph{Econometrics Journal}, 3(2), 198–215.
}
\keyword{datasets}
