\name{cpi}
\alias{cpi}
\docType{data}
\title{
Percentage Change (over 12 months) of United Kingdom Consumer Price Inflation
}
\description{
Time series of quarterly UK CPI from Q1 1997 to Q3 2013. Data formatted from the October 2013 release of the CPI data by the Office of National Statistics. Q1 are taken from February values, Q2 from May, Q3 from August and Q4 from November.  
}
\usage{data(cpi)}
\format{
  The format is:
 Time-Series [1:67] from 1997 to 2014: 88.8 89.6 90 90.4 90.3 91.5 91.2 91.7 91.5 92.7 ...
}
\source{
October 2013 CPI data by the Office of National Statistics. Retrieved from "Consumer Price Inflation Reference Tables, October 2013" spreadsheet at:

http://www.ons.gov.uk/ons/publications/re-reference-tables.html?edition=tcm%3A77-306767
}
\examples{
data(cpi)
}
\keyword{datasets}
